+++ packages/figma-utils/src/cli.ts
@@ -2,7 +2,8 @@
 import { Command } from "commander";
 import fs from "node:fs";
 import { fileURLToPath, URL } from "node:url";
-import { importCommand } from "./commands/import-variables.js";
+import { importIconsCommand } from "./commands/import-icons.js";
+import { importVariablesCommand } from "./commands/import-variables.js";
 
 const packageJson = JSON.parse(
   fs.readFileSync(fileURLToPath(new URL("../package.json", import.meta.url)), "utf8"),
@@ -11,7 +12,7 @@ const packageJson = JSON.parse(
 const cli = new Command();
 cli.version(packageJson.version, "-v, --version").description(packageJson.description);
 
-const availableCommands = [importCommand];
+const availableCommands = [importVariablesCommand, importIconsCommand];
 availableCommands.forEach((command) => cli.addCommand(command));
 
 cli.parse();