+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @sit-onyx/nuxt
 
+## 1.0.0-beta.40
+
+### Minor Changes
+
+- dfa58b8: feat: provide toasts
+
+  The Nuxt plugin now registers the `createOnyx()` Vue plugin which also provides [toasts](https://storybook.onyx.schwarz/?path=/docs/feedback-toast--docs). So you no longer need to manually provide them, so if you have something like the following in your code, you can remove it:
+
+  ```ts
+  provide(TOAST_PROVIDER_INJECTION_KEY, createToastProvider());
+  ```
+
+  Also the i18n integration that syncs your nuxt-i18n locale with onyx now also supports the `locale.language` key since `locale.iso` is deprecated.
+
+### Patch Changes
+
+- Updated dependencies [dfa58b8]
+  - sit-onyx@1.0.0-beta.39
+
+## 1.0.0-beta.39
+
+### Patch Changes
+
+- Updated dependencies [d4fe7d1]
+  - sit-onyx@1.0.0-beta.38
+
+## 1.0.0-beta.38
+
+### Patch Changes
+
+- Updated dependencies [3d612d4]
+  - sit-onyx@1.0.0-beta.37
+
+## 1.0.0-beta.37
+
+### Patch Changes
+
+- Updated dependencies [8647795]
+  - sit-onyx@1.0.0-beta.36
+
+## 1.0.0-beta.36
+
+### Patch Changes
+
+- sit-onyx@1.0.0-beta.35
+
 ## 1.0.0-beta.35
 
 ### Patch Changes